mod01-ai-services/Labfiles/04-use-a-container/stv_lab/requirements.txt
Flask
python-dotenv
azure-core
azure-ai-textanalytics
psutil
